requirements.txt
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.1
certifi==2024.6.2
charset-normalizer==3.3.2
cmake==3.29.5
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.6.0
huggingface-hub==0.23.3
idna==3.7
Jinja2==3.1.4
lightning-utilities==0.11.2
lit==18.1.7
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numpy==1.24.4
packaging==24.1
pandas==2.0.3
peft==0.11.1
protobuf==5.27.1
psutil==5.9.8
python-dateutil==2.9.0.post0
pytorch-lightning==1.8.6
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
sentencepiece==0.2.0
six==1.16.0
sympy==1.12.1
tensorboardX==2.6.2.2
tokenizers==0.13.3
torchmetrics==1.4.0.post0
tqdm==4.66.4
transformers==4.28.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
#nvidia-cublas-cu11
#nvidia-cuda-cupti-cu11
#nvidia-cuda-nvrtc-cu11
#nvidia-cuda-runtime-cu11
#nvidia-cudnn-cu11
#nvidia-cufft-cu11
#nvidia-curand-cu11
#nvidia-cusolver-cu11
#nvidia-cusparse-cu11
#nvidia-nvtx-cu11
